GRAI picking: relocate PickingJobGraiTargetService to job.service; extract GraiTuResolution with builder; nullable LU target; trim javadocs
