+++ backend/de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/picking/job/service/commands/PickingJobTestHelper.java
@@ -18,7 +18,7 @@ import de.metas.handlingunits.allocation.impl.HULoader;
 import de.metas.handlingunits.allocation.impl.HUProducerDestination;
 import de.metas.handlingunits.grai.HUGraiService;
 import de.metas.handlingunits.grai.HUPIGraiRepository;
-import de.metas.handlingunits.picking.job.service.commands.grai.PickingJobGraiTargetService;
+import de.metas.handlingunits.picking.job.service.PickingJobGraiTargetService;
 import de.metas.handlingunits.inventory.InventoryService;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_HU_PI;
